Edited the Upload action method to actually use async-await, Also in each controller's Edit action method fixed a bug that cause the uploaded file not being saved. Also added datatable JS in each controller's Index view.

+++ Controllers/ClearanceController.cs
@@ -282,10 +282,13 @@ namespace Etaa.Controllers
                             HttpContext.Session.Clear();
                             var OldFilePath = "";
                             OldFilePath = _context.Clearances.Where(f => f.ClearanceId == clearance.ClearanceId).Select(f => f.ClearanceDocumentPath).Single();
-                            FileInfo file = new FileInfo(OldFilePath);
-                            if (file.Exists)
+                            if (OldFilePath != null && !string.IsNullOrEmpty(OldFilePath))
                             {
-                                file.Delete();
+                                FileInfo file = new FileInfo(OldFilePath);
+                                if (file.Exists)
+                                {
+                                    file.Delete();
+                                }
                             }
                             clearance.ClearanceDocumentPath = NewFilePath;
                         }
